Factor out moduleName
